refactor: move bitwise/pow2 into directories with test mods
- add constants for bitwise trace accessors to core
- move test helpers to bitwise or pow2 test modules respectively
- move shared test helper to bitwise_pow2 module
