+++ air/src/aux_table/bitwise_pow2/mod.rs
@@ -6,9 +6,6 @@ use vm_core::{bitwise::NUM_SELECTORS, utils::range as create_range, Felt};
 mod bitwise;
 mod pow2;
 
-#[cfg(test)]
-mod tests;
-
 // CONSTANTS
 // ================================================================================================
 
@@ -154,3 +151,17 @@ pub const BITWISE_POW2_K1_MASK: [Felt; PERIODIC_CYCLE_LEN] = [
     Felt::ONE,
     Felt::ZERO,
 ];
+
+// TEST HELPERS
+// ================================================================================================
+
+/// Returns the values from the shared bitwise & power of two processor's periodic columns for the
+/// specified cycle row.
+#[cfg(test)]
+fn get_periodic_values(cycle_row: usize) -> [Felt; 2] {
+    match cycle_row {
+        0 => [Felt::ONE, Felt::ONE],
+        8 => [Felt::ZERO, Felt::ZERO],
+        _ => [Felt::ZERO, Felt::ONE],
+    }
+}
